+++ apps/s3n/portal/drizzle/meta/_journal.json
@@ -1,0 +1,13 @@
+{
+  "version": "7",
+  "dialect": "postgresql",
+  "entries": [
+    {
+      "idx": 0,
+      "version": "7",
+      "when": 1742457241774,
+      "tag": "0000_perpetual_thunderbolt",
+      "breakpoints": true
+    }
+  ]
+}
