+++ apps/s3n/portal/pages/billing.vue
@@ -38,4 +38,7 @@ function goto(link: string) {
       </Breadcrumb>
     </div>
   </header>
+  <main class="flex w-full h-full flex-col justify-center items-center">
+    <h1 class="text-2xl font-semibold text-center">Not available yet..</h1>
+  </main>
 </template>
