+++ apps/s3n/portal/pages/oracle/index.vue
@@ -32,10 +32,13 @@ function goto(link: string) {
           </BreadcrumbItem>
           <BreadcrumbSeparator> </BreadcrumbSeparator>
           <BreadcrumbItem>
-            <BreadcrumbPage> My Projects </BreadcrumbPage>
+            <BreadcrumbPage> Oracle </BreadcrumbPage>
           </BreadcrumbItem>
         </BreadcrumbList>
       </Breadcrumb>
     </div>
   </header>
+  <main class="flex w-full h-full flex-col justify-center items-center">
+    <h1 class="text-2xl font-semibold text-center">Not available yet..</h1>
+  </main>
 </template>
