+++ apps/s3n/portal/pages/storage/index.vue
@@ -32,10 +32,13 @@ function goto(link: string) {
           </BreadcrumbItem>
           <BreadcrumbSeparator> </BreadcrumbSeparator>
           <BreadcrumbItem>
-            <BreadcrumbPage> Schemas </BreadcrumbPage>
+            <BreadcrumbPage> Storage </BreadcrumbPage>
           </BreadcrumbItem>
         </BreadcrumbList>
       </Breadcrumb>
     </div>
   </header>
+  <main class="flex w-full h-full flex-col justify-center items-center">
+    <h1 class="text-2xl font-semibold text-center">Not available yet..</h1>
+  </main>
 </template>
